Manage tombstone lifespan.
We don't want to keep them forever or we build up state
forever. This will remove them after a timeout window.
The result is that anyone partitioned from the cluster
for more than TOMBSTONE_LIFESPAN and believing that a
service is alive will re-add it to the cluster.

Need to add functionality to flush state when hosts
leave the cluster.