Rework retransmits code and get tests passing!
This trades network bandwidth for much more aggressive
retransmitting of updates. Also fixes a bug where we were
retransmitting our own service adds on the current host.

Retransmits now happen any time we add a new service
(that isn't ours) and any time we find a service that
invalidates our own entry.

ServerChanged() no longer relies on external updates to
the server.LastUpdated, it takes the timestamp as an
argument.