Convergence improvement: Get Tombstones to retransmit.
This is not the most elegant solution. It updates the timestamp
on the Tombstone record with a 50ns additional each time for the
next 10 seconds. It's the only handle we currently have to get
other nodes to retransmit this service. If we implement a
different retransmit limiter, like a counter, we could eliminate
this. For now it's the cleanest solution.