Flexible Container Naming
This moves the container naming logic into DockerDiscovery and
injects a naming function into Service. Naming is controlled by
setting one of:

[DockerConfig]
name_from_label = set the container name to the value of the
                  container label with key matching this sting
name_match      = [ DEFAULT ] regex to use for grouping about the container
                  name.

This should be backwards compatible.