Carry IP addresses for each port
This is a really important change. It allows for hosts
to have multiple interfaces and export the containers
on the specified port. Static discovery will use any
provided ports or will use the default published IP
when this is not specified.

HAproxy will default to using the hostname as before
when the ports do not have IPs specified.

When a Docker container reports a port with an IP
address of 0.0.0.0, docker discovery will instead
replace it with the default published IP address.