Add API to annually mark services as tombstone.
This is an early WIP, sharing here to get some feedback exploring options to implement it.

Issue:
Currently we don't have a way to tell to Sidecar when a container is about to go down before it actually died. This translated into some errors when rolling out services while Sidecar detect the container as unhealthy and spread the new state in the cluster.

Goal:
Have a way for the sidecar-executor to notify Sidecar in advance when a Mesos task is going to be kill.

Design considerations:
To make it easier, I'm reusing the tombstone state but there may be some edge case I didn't consider here.
The other option is to add a new "draining" state but it'll make the service state more difficult to understand (not very familiar with the finite-state machine here)